refactor: move watcher and directory scan into the CLI layer
