+++ src/lib.rs
@@ -1,11 +1,7 @@
-use std::collections::HashMap;
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::time::Duration;
 
 use anyhow::{Context, Result, bail};
-use notify::event::{CreateKind, RemoveKind, RenameMode};
-use notify::{EventKind, RecommendedWatcher, RecursiveMode, Watcher};
 use rmcp::handler::server::tool::ToolRouter;
 use rmcp::handler::server::wrapper::Parameters;
 use rmcp::model::{LoggingLevel, LoggingMessageNotificationParam, ServerCapabilities, ServerInfo};
@@ -204,197 +200,12 @@ pub fn resolve_dir(dir: &Path) -> Result<PathBuf> {
     Ok(canonical)
 }
 
-pub fn collect_markdown_files(base_dir: &Path) -> Result<Vec<PathBuf>> {
-    let mut stack = vec![base_dir.to_path_buf()];
-    let mut files = Vec::new();
-
-    while let Some(dir) = stack.pop() {
-        for entry in fs::read_dir(&dir)
-            .with_context(|| format!("failed to read directory: {}", dir.display()))?
-        {
-            let entry = entry?;
-            let path = entry.path();
-            let file_type = entry.file_type()?;
-
-            if file_type.is_dir() {
-                stack.push(path);
-                continue;
-            }
-
-            if file_type.is_file()
-                && path
-                    .extension()
-                    .and_then(|ext| ext.to_str())
-                    .map(|ext| ext.eq_ignore_ascii_case("md"))
-                    .unwrap_or(false)
-            {
-                files.push(path);
-            }
-        }
-    }
-
-    Ok(files)
-}
-
-/// Create a `Traverze` engine and index the given Markdown files.
+/// Create a `Traverze` engine and index the given files.
 pub fn build_engine(index_dir: &Path, files: &[PathBuf]) -> Result<(Traverze, usize)> {
     let engine = Traverze::new_in_dir_for_indexing(index_dir, TokenizerMode::LinderaIpadic, true)
         .context("traverze index initialization failed")?;
     let indexed = engine
         .index_files(files)
-        .context("failed to index markdown files")?;
+        .context("failed to index files")?;
     Ok((engine, indexed))
-}
-
-/// Start watching the directory for file changes and update the index accordingly.
-///
-/// Returns the [`RecommendedWatcher`] handle. The watcher stops when this
-/// handle is dropped, so keep it alive for the lifetime of the server.
-pub fn start_watcher(engine: Traverze, base_dir: PathBuf) -> Result<RecommendedWatcher> {
-    let (tx, rx) = std::sync::mpsc::channel();
-
-    let mut watcher = RecommendedWatcher::new(
-        move |res: std::result::Result<notify::Event, notify::Error>| {
-            if let Ok(event) = res {
-                let _ = tx.send(event);
-            }
-        },
-        notify::Config::default().with_poll_interval(Duration::from_secs(2)),
-    )
-    .context("failed to create file watcher")?;
-
-    watcher
-        .watch(base_dir.as_ref(), RecursiveMode::Recursive)
-        .with_context(|| format!("failed to watch directory: {}", base_dir.display()))?;
-
-    tokio::spawn(async move {
-        tokio::task::spawn_blocking(move || {
-            let debounce = Duration::from_millis(500);
-            let mut pending: HashMap<PathBuf, EventKind> = HashMap::new();
-
-            loop {
-                // Wait for the first event (blocking).
-                let event = match rx.recv() {
-                    Ok(e) => e,
-                    Err(_) => break,
-                };
-                accumulate(&mut pending, &event);
-
-                // Drain further events until the channel is quiet for `debounce`.
-                loop {
-                    match rx.recv_timeout(debounce) {
-                        Ok(e) => accumulate(&mut pending, &e),
-                        Err(std::sync::mpsc::RecvTimeoutError::Timeout) => break,
-                        Err(std::sync::mpsc::RecvTimeoutError::Disconnected) => break,
-                    }
-                }
-
-                // Process batched events.
-                let to_index: Vec<PathBuf> = pending
-                    .iter()
-                    .filter(|(_, kind)| {
-                        matches!(kind, EventKind::Create(_) | EventKind::Modify(_))
-                    })
-                    .map(|(path, _)| path.clone())
-                    .collect();
-
-                let to_remove: Vec<PathBuf> = pending
-                    .iter()
-                    .filter(|(_, kind)| matches!(kind, EventKind::Remove(_)))
-                    .map(|(path, _)| path.clone())
-                    .collect();
-
-                pending.clear();
-
-                if !to_remove.is_empty() {
-                    match engine.remove_files(&to_remove) {
-                        Ok(n) => eprintln!("watcher: removed {} file(s) from index", n),
-                        Err(e) => eprintln!("watcher: remove error: {e}"),
-                    }
-                }
-
-                if !to_index.is_empty() {
-                    let _ = engine.remove_files(&to_index);
-                    match engine.index_files(&to_index) {
-                        Ok(n) => eprintln!("watcher: re-indexed {} file(s)", n),
-                        Err(e) => eprintln!("watcher: re-index error: {e}"),
-                    }
-                }
-            }
-        });
-    });
-
-    Ok(watcher)
-}
-
-/// Accumulate file-system events into `pending`, keeping only the latest
-/// [`EventKind`] per path and filtering to Markdown files.
-///
-/// Rename events are normalised into Remove / Create so that the processing
-/// loop does not need to know about renames.
-fn accumulate(pending: &mut HashMap<PathBuf, EventKind>, event: &notify::Event) {
-    match &event.kind {
-        EventKind::Modify(notify::event::ModifyKind::Name(mode)) => {
-            match mode {
-                // "From" carries the old path → treat as removal.
-                RenameMode::From => {
-                    if let Some(old) = event.paths.first() {
-                        if is_markdown(old) {
-                            pending.insert(old.clone(), EventKind::Remove(RemoveKind::File));
-                        }
-                    }
-                }
-                // "To" carries the new path → treat as creation.
-                RenameMode::To => {
-                    if let Some(new) = event.paths.first() {
-                        if is_markdown(new) {
-                            pending.insert(new.clone(), EventKind::Create(CreateKind::File));
-                        }
-                    }
-                }
-                // "Both" carries [old, new] in a single event.
-                RenameMode::Both => {
-                    if let Some(old) = event.paths.first() {
-                        if is_markdown(old) {
-                            pending.insert(old.clone(), EventKind::Remove(RemoveKind::File));
-                        }
-                    }
-                    if let Some(new) = event.paths.get(1) {
-                        if is_markdown(new) {
-                            pending.insert(new.clone(), EventKind::Create(CreateKind::File));
-                        }
-                    }
-                }
-                // "Any" / "Other" – direction unknown. If the file exists now
-                // treat it as a creation; otherwise as a removal.
-                _ => {
-                    for path in &event.paths {
-                        if is_markdown(path) {
-                            let kind = if path.exists() {
-                                EventKind::Create(CreateKind::File)
-                            } else {
-                                EventKind::Remove(RemoveKind::File)
-                            };
-                            pending.insert(path.clone(), kind);
-                        }
-                    }
-                }
-            }
-        }
-        // Non-rename events: pass through as-is.
-        _ => {
-            for path in &event.paths {
-                if is_markdown(path) {
-                    pending.insert(path.clone(), event.kind.clone());
-                }
-            }
-        }
-    }
-}
-
-fn is_markdown(path: &Path) -> bool {
-    path.extension()
-        .and_then(|ext| ext.to_str())
-        .map(|ext| ext.eq_ignore_ascii_case("md"))
-        .unwrap_or(false)
 }
